requirements.txt
FastMCP==1.0.0
asyncio==3.4.3